Create, export and import a module - relates #15
* Learned how to use the exports property of the module object
to export a module.

* Learned how to import a module inside a js file using the
require() method.
