+++ 01/templates/template-product.html
@@ -270,7 +270,7 @@
       <h1>🌽 Node Farm 🥦</h1>
 
       <figure class="product">
-        <div class="product__organic"><h5>Organic</h5></div>
+        <div class="product__organic {%NOT_ORGANIC%}"><h5>Organic</h5></div>
         <a href="/overview" class="product__back">
           <span class="emoji-left">👈</span>Back
         </a>
